@@ -1,16 +1,13 @@
 #!/usr/bin/env node
 "use strict";
 
-import { MF } from ".";
-import * as mri from "mri";
+import { makeFile } from ".";
+import arg from "arg";
 
-interface IArguments extends mri.Argv {
-  _: string[];
-  help: boolean;
-  h: boolean;
-}
-
-const args = <IArguments>mri(process.argv.slice(2), { boolean: ["h", "help"] });
+const args = arg({
+  "--help": Boolean,
+  "-h": "--help",
+});
 
 let help = `
   Usage
@@ -21,6 +18,13 @@ let help = `
     ./component/Button/index.js created 👍
 `;
 
-const app = () => (args.help || args.h ? console.log(help) : MF(args._));
+const app = () => {
+  if (args["--help"] || args["-h"]) {
+    console.log(help);
+    return;
+  }
+
+  makeFile(args._);
+};
 
 app();
