@@ -1,13 +1,12 @@
-import * as execa from "execa";
-import * as fs from "fs";
-import * as del from "del"
-import { MF } from "./src";
+import { execa } from "execa";
+import fs from "fs";
+import del from "del";
 
 const FILES: string[] = ["foo.js", "bar.js", "baz.js"];
 
-const LOCATION_FILES = FILES.map(f => `./dist/${f}`);
+const LOCATION_FILES = FILES.map((f) => `./dist/${f}`);
 
-describe("MF", () => {
+describe("makeFile", () => {
   it("runs and has standard out with emoji", async () => {
     const result = await execa("./dist/cli.js", LOCATION_FILES);
     expect(result.stdout).toContain("🌈");
@@ -17,7 +16,7 @@ describe("MF", () => {
   it("produces files", () => {
     const files = fs.readdirSync("./dist");
 
-    FILES.forEach(f => expect(files).toContain(f));
+    FILES.forEach((f) => expect(files).toContain(f));
   });
 
   afterAll(() => del(LOCATION_FILES));
